Help
I have been unable to giet the code to run. The Caliberate is the simplest, yet the issue seems to  be in the actual i2c operations in MPU9250, or MPU6500. The i2c devices
are found with the scan function. When I put an ossilscope on the I2C lines, there is no activity. I get "TypeError: "property"  object isn't subscriptable" It appears
the returned "value" is the string "property" rather than a number. Since ther is no activity on the I2C lines, that makes sense. 
I have spent quit a bit of time, but am stumped at the moment. I have the PCB built, the body and parts printed and assembled, along with the steppers.
Any help  would be appreciated.   Thanks